+++ muffinscript/__init__.py
@@ -1,0 +1,16 @@
+from muffinscript.error import (
+    ProgramSyntaxError,
+    output_error,
+)
+from muffinscript.interpreter import evaluate_expression
+from muffinscript.lexer import tokenize
+from muffinscript.parser import parse_tokens
+
+
+__all__ = [
+    'ProgramSyntaxError',
+    'output_error',
+    'evaluate_expression',
+    'tokenize',
+    'parse_tokens',
+]